curve test done

+++ contracts/exchanges/interfaces/ICurvePool.sol
@@ -3,8 +3,8 @@ pragma solidity ^0.8.0;
 
 interface ICurvePool {
     function exchange_underlying(
-        int128 i,
-        int128 j,
+        uint256 i,
+        uint256 j,
         uint256 dx,
         uint256 min_dy
     ) external;
@@ -12,8 +12,8 @@ interface ICurvePool {
 
 interface ICurvePoolToETH {
     function exchange(
-        int128 i,
-        int128 j,
+        uint256 i,
+        uint256 j,
         uint256 dx,
         uint256 min_dy,
         bool use_eth
